refactor: extract applyTextChanges
