+++ src/applyCodeFix.ts
@@ -1,34 +1,11 @@
 import ts from 'typescript';
 import { FileService } from './FileService.js';
+import { applyTextChanges } from './util/applyTextChanges.js';
 
 export const fixIdDelete = 'unusedIdentifier_delete';
 export const fixIdDeleteImports = 'unusedIdentifier_deleteImports';
 
 type FixId = typeof fixIdDelete | typeof fixIdDeleteImports;
-
-export const applyTextChanges = (
-  oldContent: string,
-  changes: readonly ts.TextChange[],
-) => {
-  const result: string[] = [];
-
-  const sortedChanges = [...changes].sort(
-    (a, b) => a.span.start - b.span.start,
-  );
-
-  let currentPos = 0;
-
-  for (const change of sortedChanges) {
-    result.push(oldContent.slice(currentPos, change.span.start));
-    result.push(change.newText);
-
-    currentPos = change.span.start + change.span.length;
-  }
-
-  result.push(oldContent.slice(currentPos));
-
-  return result.join('');
-};
 
 export const applyCodeFix = ({
   fixId,
